move Program execution into worker
